Adding meal view fixed

+++ src/main/java/de/seibushin/bodyArchitect/model/nutrition/Food.java
@@ -145,7 +145,7 @@ public class Food {
         return portion;
     }
 
-    public void setPortion(int portion) {
+    public void setPortion(double portion) {
         this.portion = portion;
     }
